Standardise download specification

+++ src/AUMCdb_MEDS/download.py
@@ -11,6 +11,7 @@ from .commands import run_command
 def download_data(
     output_dir: Path,
     dataset_info: DictConfig,
+    do_demo: bool = False,
     runner_fn: callable = run_command,
 ):
     """Downloads the data specified in dataset_info.dataset_urls to the output_dir.
@@ -18,7 +19,8 @@ def download_data(
     Args:
         output_dir: The directory to download the data to.
         dataset_info: The dataset information containing the URLs to download.
-        do_demo: If True, download the demo URLs instead of the main URLs.
+        do_demo: Not currently available for AUMCdb; retained for compatibility
+            but will raise an error if True.
         runner_fn: The function to run the command with (added for dependency injection).
 
     Raises:
@@ -26,42 +28,52 @@ def download_data(
 
     Examples:
         >>> cfg = DictConfig({
-        ...     "dataset_url": "http://example.com/dataset",
-        ...     "api_key": "[MY_API_KEY]"
+        ...     "urls": {
+        ...         "dataset": [{
+        ...             "url": "http://example.com/dataset",
+        ...             "api_key": "abcdefg123"
+        ...         }]
+        ...     }
         ... })
         >>> def fake_shell_succeed(cmd):
         ...     print(" ".join(cmd))
-        >>> def fake_shell_fail(cmd):
-        ...     raise ValueError(f"Failed to run {' '.join(cmd)}")
         >>> download_data(Path("data"), cfg, runner_fn=fake_shell_succeed)
-        curl -L -o data/AUMCdb.zip -H X-Dataverse-key:[MY_API_KEY] http://example.com/dataset
+        curl -L -o data/AUMCdb.zip -H X-Dataverse-key:abcdefg123 http://example.com/dataset
     """
 
-    url = dataset_info.get("dataset_url", None)
-    if url is None:
-        url = getpass("Enter the download link: ")
+    if do_demo:
+        raise ValueError("Demo download is not currently available for AUMCdb.")
+    else:
+        urls = dataset_info.get("urls", [])  # .get(["numeric_items","dataset"], [])
 
-    output_file = output_dir / "AUMCdb.zip"
+        for i, entry in enumerate(urls):
+            url = urls[entry][0].get("url", None)
+            if url is None:
+                url = input("Enter the download link: ")
 
-    if output_file.exists():
-        logging.info(f"Removing existing file {output_file}")
-        output_file.unlink()
+            key = urls[entry][0].get("api_key", None)
+            if key is None:
+                key = getpass("Enter your API Token: ")
 
-    key = dataset_info.get("api_key", None)
-    if key is None:
-        key = getpass("Enter your API Token: ")
+            output_file = urls[entry][0].get("output_file", None)
 
-    command_parts = ["curl", "-L", "-o", str(output_file), "-H", f"X-Dataverse-key:{key}", url]
+            # output_file = output_dir #/ f"AUMCdb_{i}"
 
-    try:
-        runner_fn(command_parts)
-    except ValueError as e:
-        raise ValueError(f"Failed to download data from {url}") from e
+            if (output_dir / output_file).exists():
+                logging.info(f"Removing existing file {output_file}")
+                output_file.unlink()
+            command_parts = ["cd", str(output_dir), "&&"]
+            command_parts.extend(["curl", "-L", "-O", "-J", "-H", f"X-Dataverse-key:{key}", url])
 
-    with zipfile.ZipFile(output_file, "r") as zip_ref:
-        zip_ref.extractall(output_dir)
-    logging.info(f"Downloaded and extracted data to {output_dir}")
+            try:
+                runner_fn(command_parts)
+            except ValueError as e:
+                raise ValueError(f"Failed to download data from {url}") from e
 
-    if output_file.exists():
-        logging.info(f"Removing existing file {output_file}")
-        output_file.unlink()
+            with zipfile.ZipFile(output_file, "r") as zip_ref:
+                zip_ref.extractall(output_dir)
+            logging.info(f"Downloaded and extracted data to {output_dir}")
+
+            if output_file.exists():
+                logging.info(f"Removing zipped file {output_file}")
+                output_file.unlink()
